# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Система контроля посещаемости и успеваемости",
    "description": "REST API ВСПК: занятия, посещаемость по свайпам/QR, оценки и отчёты.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Регистрация пользователя",
        "description": "Создаёт пользователя. Для готового студенческого профиля используйте тестовые данные seed.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Register Api V1 Auth Register Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Авторизация",
        "description": "Принимает JSON email/password или OAuth2 form username/password. Возвращает JWT.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "properties": {
                  "username": {
                    "type": "string",
                    "format": "email",
                    "example": "teacher@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Teacher123!"
                  }
                },
                "type": "object",
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Обновить JWT",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Текущий профиль",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/attendance/swipe/lessons/{lesson_id}/swipe-queue": {
      "get": {
        "tags": [
          "Посещаемость — Свайп"
        ],
        "summary": "Очередь карточек для свайпа",
        "description": "Получить очередь неотмеченных студентов (ФИО, фото, ID) для последовательного свайпа.",
        "operationId": "get_swipe_queue_api_v1_attendance_swipe_lessons__lesson_id__swipe_queue_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwipeQueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/swipe/lessons/{lesson_id}/swipe": {
      "post": {
        "tags": [
          "Посещаемость — Свайп"
        ],
        "summary": "Отправить результат свайпа",
        "description": "Отметить студента как PRESENT или ABSENT. Возвращает следующего студента и количество оставшихся.",
        "operationId": "process_swipe_api_v1_attendance_swipe_lessons__lesson_id__swipe_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwipeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwipeResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/grades/lessons/{lesson_id}/grades": {
      "post": {
        "tags": [
          "Оценки"
        ],
        "summary": "Выставить оценку",
        "description": "Оценка 2–5 доступна только присутствующему или опоздавшему студенту.",
        "operationId": "create_grade_api_v1_grades_lessons__lesson_id__grades_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Оценки"
        ],
        "summary": "Оценки за занятие",
        "operationId": "get_grades_api_v1_grades_lessons__lesson_id__grades_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GradeWithStudentName"
                  },
                  "title": "Response Get Grades Api V1 Grades Lessons  Lesson Id  Grades Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/lessons/{lesson_id}/summary": {
      "get": {
        "tags": [
          "Отчёты"
        ],
        "summary": "Итоговая статистика занятия",
        "description": "Процент посетивших, средний балл и детализация по студентам.",
        "operationId": "lesson_summary_api_v1_reports_lessons__lesson_id__summary_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/export/excel": {
      "get": {
        "tags": [
          "Отчёты"
        ],
        "summary": "Экспорт ведомости в CSV",
        "description": "CSV в UTF-8 с BOM открывается в Excel. Передайте lesson_id.",
        "operationId": "export_excel_api_v1_reports_export_excel_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/csv": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttendanceStatus": {
        "type": "string",
        "enum": [
          "PRESENT",
          "ABSENT",
          "LATE"
        ],
        "title": "AttendanceStatus"
      },
      "AttendanceSummary": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "status": {
            "$ref": "#/components/schemas/AttendanceStatus"
          },
          "method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttendanceMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "marked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marked At"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "full_name",
          "status"
        ],
        "title": "AttendanceSummary"
      },
      "GradeCreate": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id",
            "description": "ID записи студента в таблице students"
          },
          "grade_value": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 2.0,
            "title": "Grade Value",
            "description": "Оценка: 2, 3, 4 или 5"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Комментарий к оценке"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "grade_value"
        ],
        "title": "GradeCreate"
      },
      "GradeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "grade_value": {
            "type": "integer",
            "title": "Grade Value"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lesson_id",
          "student_id",
          "grade_value",
          "created_at"
        ],
        "title": "GradeResponse"
      },
      "GradeWithStudentName": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "grade_value": {
            "type": "integer",
            "title": "Grade Value"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lesson_id",
          "student_id",
          "grade_value",
          "created_at",
          "full_name"
        ],
        "title": "GradeWithStudentName"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LessonSummaryResponse": {
        "properties": {
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "total_students": {
            "type": "integer",
            "title": "Total Students"
          },
          "present_count": {
            "type": "integer",
            "title": "Present Count"
          },
          "absent_count": {
            "type": "integer",
            "title": "Absent Count"
          },
          "late_count": {
            "type": "integer",
            "title": "Late Count"
          },
          "attendance_percent": {
            "type": "number",
            "title": "Attendance Percent"
          },
          "average_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Grade"
          },
          "attendance": {
            "items": {
              "$ref": "#/components/schemas/AttendanceSummary"
            },
            "type": "array",
            "title": "Attendance"
          },
          "grades": {
            "items": {
              "$ref": "#/components/schemas/GradeWithStudentName"
            },
            "type": "array",
            "title": "Grades"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "subject_name",
          "total_students",
          "present_count",
          "absent_count",
          "late_count",
          "attendance_percent",
          "attendance",
          "grades"
        ],
        "title": "LessonSummaryResponse"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest"
      },
      "Role": {
        "type": "string",
        "enum": [
          "TEACHER",
          "STUDENT",
          "ADMIN"
        ],
        "title": "Role"
      },
...
      "SwipeQueueResponse": {
        "properties": {
          "students": {
            "items": {
              "$ref": "#/components/schemas/SwipeStudentCard"
            },
            "type": "array",
            "title": "Students"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "students",
          "total_count"
        ],
        "title": "SwipeQueueResponse"
      },
      "SwipeRequest": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id",
            "description": "ID записи студента в таблице students"
          },
          "status": {
            "$ref": "#/components/schemas/AttendanceStatus",
            "description": "PRESENT или ABSENT"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "status"
        ],
        "title": "SwipeRequest"
      },
      "SwipeResultResponse": {
        "properties": {
          "next_student": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SwipeStudentCard"
              },
              {
                "type": "null"
              }
            ]
          },
          "remaining_count": {
            "type": "integer",
            "title": "Remaining Count"
          }
        },
        "type": "object",
        "required": [
          "remaining_count"
        ],
        "title": "SwipeResultResponse"
      },
      "SwipeStudentCard": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "student_card_number": {
            "type": "string",
            "title": "Student Card Number"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "full_name",
          "student_card_number"
        ],
        "title": "SwipeStudentCard"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse"
      },
      "UserLoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "id",
          "full_name",
          "email",
          "role"
        ],
        "title": "UserLoginResponse"
      },
      "UserMeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "email",
          "role",
          "created_at"
        ],
        "title": "UserMeResponse"
      },
      "UserRegister": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Full Name",
            "description": "ФИО пользователя",
            "default": "Гость"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email пользователя"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password",
            "description": "Пароль"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Роль: TEACHER, STUDENT, ADMIN",
            "default": "STUDENT"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserRegister"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}